refactor: ユーティリティ関数を整理し、関連するモジュールに移動
